Email Encoding
If the actual email address is in the html file, it is at risk of being found and used by spambots. To avoid this, sometimes an email encoder like this one http://wbwip.com/wbw/emailencoder.html is used to encrypt the address. I have copy pasted the code where I saw the address. It should display and link like a normal email address on the website, but not be as vulnerable to spambots.